fet: no error

+++ houtu-scene/src/math/epsilon.rs
@@ -50,7 +50,7 @@ pub const LUNAR_RADIUS: f64 = 1737400.0;
 
 pub const SIXTY_FOUR_KILOBYTES: u32 = 64 * 1024;
 
-pub const FOUR_GIGABYTES: u32 = 4 * 1024 * 1024 * 1024;
+pub const FOUR_GIGABYTES: u64 = 4 * 1024 * 1024 * 1024;
 
 pub const RADIANS_PER_DEGREE: f64 = PI / 180.0;
